Per ingredient sync table
