+++ src/input/input_field.rs
@@ -1,7 +1,8 @@
 use crate::accumulator::accumulated_map::{AccumulatedMap, InputAccumulatedValues};
 use crate::cycle::CycleRecoveryStrategy;
 use crate::ingredient::{fmt_index, Ingredient};
-use crate::input::{Configuration, MemoTableTypes};
+use crate::input::Configuration;
+use crate::table::memo::MemoTableTypes;
 use crate::zalsa::IngredientIndex;
 use crate::zalsa_local::QueryOrigin;
 use crate::{Database, DatabaseKeyIndex, Id, Revision};
