+++ src/tracked_struct.rs
@@ -13,7 +13,6 @@ use crate::{
     salsa_struct::SalsaStructInDb,
     table::{
         memo::{MemoTable, MemoTableTypes},
-        sync::SyncTable,
         Slot, Table,
     },
     zalsa::{IngredientIndex, Zalsa},
@@ -231,9 +230,6 @@ where
 
     /// Memo table storing the results of query functions etc.
     memos: MemoTable,
-
-    /// Sync table storing the results of query functions etc.
-    syncs: SyncTable,
 }
 // ANCHOR_END: ValueStruct
 
@@ -333,7 +329,6 @@ where
             fields: unsafe { self.to_static(fields) },
             revisions: C::new_revisions(current_deps.changed_at),
             memos: Default::default(),
-            syncs: Default::default(),
         };
 
         if let Some(id) = self.free_list.pop() {
@@ -725,14 +720,6 @@ where
         // when deleting a tracked struct.
         self.read_lock(current_revision);
         &self.memos
-    }
-
-    unsafe fn syncs(&self, current_revision: Revision) -> &crate::table::sync::SyncTable {
-        // Acquiring the read lock here with the current revision
-        // ensures that there is no danger of a race
-        // when deleting a tracked struct.
-        self.read_lock(current_revision);
-        &self.syncs
     }
 
     unsafe fn drop_memos(&mut self, types: &MemoTableTypes) {
